claude-backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
waitress==3.0.2
requests==2.31.0
python-dotenv==1.0.0
anthropic>=0.39.0
openai>=1.40.0
google-genai>=1.0.0
websocket-client>=1.6.0
pyyaml>=6.0
PyPDF2>=3.0.0
python-docx>=1.0.0
python-telegram-bot>=20.0
twilio>=8.10.0
discord.py>=2.4.0
httpx>=0.24.0
mcp>=1.0.0
edge-tts>=6.1.0
gemini_webapi>=1.21.0
curl_cffi>=0.7.0
